switch to custom page div
CSS doesn't have any way of styling elements based on whether they are
on the left or right page, so the planner cannot be styled using
`@page` rules. Use a custom `page` `div` instead, with classes for
left and right pages. Update the cover, title page, entries, and
numbers sections.